--- src/main/scala/net/liftweb/widgets/flot/FlotUtils.scala
@@ -1,27 +1,0 @@
-/*
- * Copyright 2007-2010 WorldWide Conferencing, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.liftweb {
-package widgets {
-package flot {
-
-case class FlotInfo (idPlaceholder: String, series: List[FlotSerie], options: FlotOptions)
-
-case class FlotNewData (series: List [FlotSerie], datas: List[(Double, Double)])
-
-}
-}
-}
