+++ src/test/scala/webapptest/snippet/AutoCompleteDemo.scala
@@ -21,9 +21,11 @@ import _root_.scala.xml.{NodeSeq, Text}
 import _root_.net.liftweb.http._
 import _root_.net.liftweb.common._
 import _root_.net.liftweb.util._
-import _root_.net.liftweb.widgets.autocomplete._
 import _root_.net.liftweb.util.Helpers._
 import _root_.net.liftweb.http.{S, RequestVar}
+
+import _root_.net.liftmodules.widgets.autocomplete._
+
 
 object posted extends RequestVar[Box[String]](Empty)
