+++ src/test/scala/webapptest/snippet/CalendarMonthViewDemo.scala
@@ -30,7 +30,7 @@ import _root_.net.liftweb.http.js._
 import _root_.java.util.Calendar
 import _root_.java.util.Calendar._
 
-import _root_.net.liftweb.widgets.calendars._
+import _root_.net.liftmodules.widgets.calendars._
 
 
 class CalendarMonthViewDemo {
