+++ src/test/scala/webapptest/snippet/TableSorterDemo.scala
@@ -20,7 +20,7 @@ package snippet {
 import _root_.scala.xml.{NodeSeq, Text}
 import _root_.net.liftweb.http.SHtml._
 import _root_.net.liftweb.util.Helpers._
-import _root_.net.liftweb.widgets.tablesorter.{TableSorter, DisableSorting, Sorting, Sorter}
+import _root_.net.liftmodules.widgets.tablesorter.{TableSorter, DisableSorting, Sorting, Sorter}
 
 class TableSorterDemo {
